+++ packages/theme/src/prebuilt/js/design-token-fallbacks.mjs
@@ -56,6 +56,7 @@ export default {
 		'var(--wp-admin-theme-color, #3858e9)',
 	'--wpds-color-background-thumb-brand-active':
 		'var(--wp-admin-theme-color, #3858e9)',
+	'--wpds-color-background-thumb-brand-disabled': '#dbdbdb',
 	'--wpds-color-background-thumb-neutral-disabled': '#dbdbdb',
 	'--wpds-color-background-thumb-neutral-weak': '#8d8d8d',
 	'--wpds-color-background-thumb-neutral-weak-active': '#6e6e6e',